+++ src/MakeItEasy/Internal/SubjectBuilder.cs
@@ -10,7 +10,8 @@ namespace MakeItEasy.Internal
     {
         public static T BuildSubject<T>(Type[] argumentTypes, object?[] argumentValues, Type[] collaboratorTypes, out object[] collaboratatorValues)
         {
-            var exceptions = new List<ArgumentCreationException>();
+            var argumentExceptions = new List<ArgumentCreationException>();
+            var constructorExceptions = new List<ConstructorFailedException>();
 
             foreach (var candidateConstructor in GetAllConstructors(typeof(T)).Select(c => SubjectConstructor<T>.TryCreateFrom(c, argumentTypes, collaboratorTypes)).Where(c => c is not null))
             {
@@ -21,26 +22,51 @@ namespace MakeItEasy.Internal
                 }
                 catch (ArgumentCreationException e)
                 {
-                    exceptions.Add(e);
+                    argumentExceptions.Add(e);
+                }
+                catch (ConstructorFailedException e)
+                {
+                    constructorExceptions.Add(e);
                 }
             }
 
-            if (exceptions.Count == 0)
+            if (argumentExceptions.Count == 0 && constructorExceptions.Count == 0)
             {
                 var requiredTypes = argumentTypes.Concat(collaboratorTypes).ToList();
                 throw new CreationException(ExceptionMessages.NoAccessibleConstructor(typeof(T), requiredTypes));
             }
 
             var message = new StringBuilder()
-                .AppendLine(ExceptionMessages.UnableToCreateSubject(typeof(T)))
-                .AppendLine()
-                .AppendLine("  " + ExceptionMessages.UnableToCreateArgument());
-            foreach (var exception in exceptions)
+                .AppendLine(ExceptionMessages.UnableToCreateSubject(typeof(T)));
+
+            if (argumentExceptions.Count > 0)
             {
                 message
-                    .Append("    (")
-                    .Append(string.Join(", ", exception.Constructor.GetParameters().Select((p, i) => i == exception.ParameterIndex ? "*" + p.ParameterType : p.ParameterType.ToString())))
-                    .AppendLine(")");
+                    .AppendLine()
+                    .AppendLine("  " + ExceptionMessages.UnableToCreateArgument());
+                foreach (var exception in argumentExceptions)
+                {
+                    message
+                        .Append("    (")
+                        .Append(string.Join(", ", exception.Constructor.GetParameters().Select((p, i) => i == exception.ParameterIndex ? "*" + p.ParameterType : p.ParameterType.ToString())))
+                        .AppendLine(")");
+                }
+            }
+
+            if (constructorExceptions.Count > 0)
+            {
+                message
+                    .AppendLine()
+                    .AppendLine("  " + ExceptionMessages.ConstructorFailed());
+                foreach (var exception in constructorExceptions)
+                {
+                    message
+                        .Append("    (")
+                        .Append(string.Join(", ", exception.Constructor.GetParameters().Select(p => p.ParameterType.ToString())))
+                        .AppendLine(")")
+                        .AppendLine("      " + ExceptionMessages.ExceptionType(exception.InnerException!))
+                        .AppendLine("      " + ExceptionMessages.ExceptionMessage(exception.InnerException!));
+                }
             }
 
             throw new CreationException(message.ToString());