+++ tests/SpecDefinitions/TestTypes/ICannotBeDummied.cs
@@ -1,0 +1,6 @@
+namespace MakeItEasy.Specs.TestTypes
+{
+    public interface ICannotBeDummied
+    {
+    }
+}